fix(c8yscrn): Fix setting of screenshot target folder overwrite  (#165)
The screenshot folder was cleared before every run of `c8yscrn`. To
clear the screenshot folder, use the `--clear` command line option when
running `c8yscrn` to avoid unintented data loss. The `global.overwrite`
setting in the workflow file is now considered for `run`. If disabled,
existing screenshots are not overwritten.