fix(c8yscrn): Allow skipping screenshots when not different (#186)
Diffing now allows comparing an possibly existing screenshot in the target folder with the new screenshot created. If a screenshot is different, a diff images with a change mask can be saved. If screenshots are equal, updating the screenshot in the target location can be skipped. See cli options and Screenshot Automation doc for more infos.