fix(commands): Fixed showCookieBanner and added disableCookieBanner (#213)
The `cy.showCookieBanner` command did remove the accepted cookie from the session, but the public options request was still intercepted and would set a new cookie. With this fix, the interception will not set a new cookie. To address possible issues with `policyVersion` now required to match for the cookie banner to not show, the `cy.disableCookieBanner` command has been introduced. By disabling the cookie banner, Cumulocity
will not check cookie and policy versions.