fix: cy.highlight style for elements with rounded corners (#613)
When using `cy.highlight()` for elements with rounded corners style, the highlight border and style would use rounded corners. Now a detached highlight is used in this case to make the highlight style independent of rounded corners of the highlighted element. Its now also possible to defined `detached` attribute for highlight in yaml workflows to overwrite the default behaviour.